fix: files/expected-array.json file
The provided file `files/expected-array.json` which contains the scraped
data in JSON format had an error, because it was missing curly braces to
define a JavaScript object.

I've added the missing curly braces in the first and last line of the
file and didn't indent the file, because I didn't want to add all that
noise in a git commit.